Update shell extension
- fix(filetree): restructure filetree to adhere to solid principles
- Initial commit
