+++ extensions/shell/CHANGELOG.md
@@ -1,38 +1,49 @@
 # shell Changelog
 
+## [Restructure filetree] 2024-09-29
+
+Break down main `index.tsx` into smaller components, helpers and custom hooks.
+
 ## [Add result indication] 2024-07-26
+
 Fix partially [issue](https://github.com/raycast/extensions/issues/13496), displaying an indicator on command finishes, with success or failure.
 
-
 ## [Fix] 2024-07-18
-Fix [issue](https://github.com/raycast/extensions/issues/13015) to avoid UI hanging on large data from `child_process`. 
+
+Fix [issue](https://github.com/raycast/extensions/issues/13015) to avoid UI hanging on large data from `child_process`.
 
 ## [Fix] 2024-05-15
+
 Fix an [issue](https://github.com/raycast/extensions/issues/12350) where commands aren't send to iTerm2.
 
-
 ## [Add alpha kitty support] 2024-05-05
-This version adds the option to run commands in `kitty` terminal. 
+
+This version adds the option to run commands in `kitty` terminal.
 Checkout the preferences and select your favorite console application in the dropdown to get going.
 
 ## [Fix] 2024-04-28
+
 Allow to run alias commands while running the shell command in raycast.
 
-
 ## [Fix] 2024-04-26
+
 Fix an issue where iTerm opens two windows instead of one when iTerm was not already running.
 
 ## [Fixes] 2023-01-02
+
 Fix an issue when the use Terminal/iTerm option is turned-on and command run with argument by moving execution logic to useEffect.
 
 ## [Add Warp Support] - 2022-11-18
-This version adds the option to run commands in Warp. 
+
+This version adds the option to run commands in Warp.
 Checkout the preferences and select your favorite console application in the dropdown to get going.
 
 ## [Update] - 2022-09-10
+
 Open new iTerm tab if iTerm is running and chooses Execute in iTerm.app.
 
 ## [iTerm fixes] - 2022-06-30
+
 fix an issue, where commands did not run, when iTerm was not open.
 
 ## [Initial Version] - 2022-06-21
