+++ extensions/shell/package.json
@@ -8,7 +8,8 @@
   "contributors": [
     "emiara",
     "kevin_manca",
-    "mohamedk1"
+    "mohamedk1",
+    "ronny011-dy"
   ],
   "categories": ["Developer Tools"],
   "license": "MIT",
