+++ extensions/shell/src/components/Command/components/ActionList/index.ts
@@ -1,0 +1,1 @@
+export { ActionList } from "./ActionList";
