+++ extensions/shell/src/components/Command/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Command";
